Tie services to Angular DI system
The `ApplicationinsightsAngularpluginErrorService` has been storing its instance
on its static property called `instance`. This is an anti-pattern because it breaks
"replugging" for microfrontends and server-side rendering. Given multiple Angular apps
that share the same dependency at runtime. The second bootstrapped Angular application
will overwrite the `instance` property (which was firstly created by the first bootstrapped
Angular application). This also means the first Angular application will be using wrong
instance of the `ApplicationinsightsAngularpluginErrorService`.

This commit also makes `AngularPlugin` a dependency injection "unit". Since it's decorated
with `@Injectable` it's now possible to "mock" the class in unit tests and developers don't
have to do any initializations manually (e.g., `new AngularPlugin()`). It's enough to inject
the `AngularPlugin` in any constructor.

Note: this is considered as a breaking change.