Run async stuff outside Angular
This commit wraps the tracking functionality to run outside of the Angular zone. Previously, it hindered
server-side rendering and hydration, causing instability in the app. The app achieves stability when no
microtasks or macrotasks are running.

On the client side, this change also prevents unnecessary view updates when asynchronous tasks are set
up by `trackPageView`.